+++ Fun.Build/StageContextExtensions.fs
@@ -16,7 +16,8 @@ module StageContextExtensionsInternal =
             Name = name
             IsActive = fun _ -> true
             IsParallel = fun _ -> false
-            ContinueOnStepFailure = false
+            ContinueStepOnStepFailure = false
+            ContinueStageOnStepFailure = false
             Timeout = ValueNone
             TimeoutForStep = ValueNone
             WorkingDir = ValueNone
@@ -250,7 +251,8 @@ module StageContextExtensionsInternal =
                                 )
                                 if i = stage.Steps.Length - 1 then AnsiConsole.WriteLine()
 
-                                if not isSuccess && not stage.ContinueOnStepFailure then stepErrorCTS.Cancel()
+                                if not isSuccess && not (stage.ContinueStepOnStepFailure || stage.ContinueStageOnStepFailure) then
+                                    stepErrorCTS.Cancel()
                                 return isSuccess
 
                             with
@@ -266,7 +268,8 @@ module StageContextExtensionsInternal =
                                 AnsiConsole.MarkupLineInterpolated $"[red]{prefix} exception hanppened.[/]"
                                 AnsiConsole.WriteException ex
                                 stepExns.Add ex
-                                if not stage.ContinueOnStepFailure then stepErrorCTS.Cancel()
+                                if not (stage.ContinueStepOnStepFailure || stage.ContinueStageOnStepFailure) then
+                                    stepErrorCTS.Cancel()
                                 return false
                         })
 
@@ -281,7 +284,7 @@ module StageContextExtensionsInternal =
                                         completers.Add completer
 
                                     let mutable i = 0
-                                    while i < completers.Count && (stage.ContinueOnStepFailure || isSuccess) do
+                                    while i < completers.Count && (stage.ContinueStepOnStepFailure || stage.ContinueStageOnStepFailure || isSuccess) do
                                         let! result = completers[i]
                                         i <- i + 1
                                         isSuccess <- isSuccess && result
@@ -290,7 +293,7 @@ module StageContextExtensionsInternal =
                                 async {
                                     let mutable i = 0
                                     let length = Seq.length steps
-                                    while i < length && (stage.ContinueOnStepFailure || isSuccess) do
+                                    while i < length && (stage.ContinueStepOnStepFailure || stage.ContinueStageOnStepFailure || isSuccess) do
                                         let! completer = Async.StartChild(Seq.item i steps, timeoutForStep)
                                         let! result = completer
                                         i <- i + 1
@@ -339,7 +342,7 @@ module StageContextExtensionsInternal =
             finally
                 pipeline |> Option.iter (fun x -> x.RunAfterEachStage stage)
 
-            stage.ContinueOnStepFailure || isSuccess, stepExns
+            stage.ContinueStageOnStepFailure || isSuccess, stepExns
 
 
     let inline buildStageIsActive ([<InlineIfLambda>] build: BuildStage) ([<InlineIfLambda>] conditionFn) =