+++ Fun.Build/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Support continueStepOnStepFailure and continueStageOnStepFailure
+
 ## [1.1.0] - 2024-01-31
 
 - Support continueOnStepFailure for stage